refactor(simulation): split preview smoke responsibilities
Separate qualification inputs, shared validation, SKY130 validation, and hosted run orchestration while preserving the preview smoke CLI and its 28 public helper exports. This keeps the numerical acceptance scenarios unchanged and makes each layer independently reviewable.

Validation: pnpm test:local scripts/preview-simulation-smoke.test.mjs scripts/lib/simulation-run-evidence.test.mjs scripts/preview-workflow.test.mjs (37 passed); node --check on the facade and all extracted modules; public export comparison against origin/main (28/28, no delta); git diff --check.

Test-Impact: no-test-change — existing preview smoke, workflow, and evidence tests exercise the unchanged public contracts after extraction
